+++ order-service/src/main/java/com/rushcrew/order_service/infrastructure/dto/point/PointDeductResponse.java
@@ -1,7 +1,5 @@
 package com.rushcrew.order_service.infrastructure.dto.point;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class PointDeductResponse {
 	private boolean success;
 	private String message;
-	private BigDecimal remainingPoint; // 차감 후 남은 포인트
+	private Long remainingPoint; // 차감 후 남은 포인트
 }
